# api/openapi/zke-server.v1.yaml
openapi: 3.1.0
info:
  title: ZKE Server API
  version: v1
  description: |
    ZKE 用户管理、资源管理、Cluster 接入管理和内部 Agent 注册使用的 HTTP API 契约。
    所有 JSON 时间均为 RFC 3339，并由当前 API 以 UTC+8 偏移返回。
servers:
  - url: /
    description: Console 与 API 使用同源入口；本地开发通过前端代理转发
...
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      operationId: listUsers
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListOffset"
        - $ref: "#/components/parameters/ListSearch"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, locked, disabled]
      responses:
        "200":
          description: 用户列表
          content:
            application/json:
              schema:
                type: object
                required: [users, pagination]
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManagedUser"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      operationId: createUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: 用户已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/v1/users/{user_id}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      tags: [Users]
      operationId: getUser
      security:
        - sessionCookie: []
      responses:
        "200":
          description: 用户详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Users]
      operationId: updateUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: 用户资料已更新
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [Users]
      operationId: deleteUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirmation"
      responses:
        "200":
          description: 用户已逻辑删除并撤销全部 Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{user_id}/status:
    parameters:
      - $ref: "#/components/parameters/UserID"
    put:
      tags: [Users]
      operationId: setUserStatus
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserStatusRequest"
      responses:
        "200":
          description: 用户状态已更新，禁用会撤销现有 Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{user_id}/unlock:
    parameters:
      - $ref: "#/components/parameters/UserID"
    post:
      tags: [Users]
      operationId: unlockUser
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirmation"
      responses:
        "200":
          description: 锁定计数已清除，账户恢复为 active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/users/{user_id}/password-reset:
    parameters:
      - $ref: "#/components/parameters/UserID"
    post:
      tags: [Users]
      operationId: resetUserPassword
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: 密码已重置，目标用户全部 Session 已撤销
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/role-bindings:
    get:
      tags: [RBAC]
      operationId: listRoleBindings
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ListLimit"
        - $ref: "#/components/parameters/ListOffset"
        - $ref: "#/components/parameters/ListSearch"
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, viewer]
        - name: scope_type
          in: query
          schema:
            type: string
            enum: [global, tenant, project]
      responses:
        "200":
          description: RoleBinding 列表
          content:
            application/json:
              schema:
                type: object
                required: [role_bindings, pagination]
                properties:
                  role_bindings:
                    type: array
                    items:
                      $ref: "#/components/schemas/RoleBinding"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [RBAC]
      operationId: createRoleBinding
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleBindingRequest"
      responses:
        "201":
          description: RoleBinding 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBinding"
        "200":
          description: 相同 RoleBinding 已存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBinding"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/role-bindings/{role_binding_id}:
    get:
      tags: [RBAC]
      operationId: getRoleBinding
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RoleBindingID"
      responses:
        "200":
          description: RoleBinding 详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBinding"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [RBAC]
      operationId: deleteRoleBinding
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/RoleBindingID"
        - $ref: "#/components/parameters/CSRFToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Confirmation"
      responses:
        "204":
          description: RoleBinding 已删除
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: zke_session
    enrollmentToken:
      type: http
      scheme: bearer
      bearerFormat: Agent enrollment token
    installationToken:
      type: http
      scheme: bearer
      bearerFormat: Agent installation token
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
...
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
    RoleBindingID:
      name: role_binding_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UUID"
...
    ListLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    ListOffset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 1000000
        default: 0
    ListSearch:
      name: q
      in: query
      schema:
        type: string
        maxLength: 253
...
  responses:
    InvalidRequest:
      description: 输入无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: 未认证或 Session 已失效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: CSRF 或权限检查失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 目标不存在或不可见
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: 幂等、生命周期或安全不变量冲突
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
      examples:
        - "2026-07-26T14:30:00+08:00"
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
...
    UserIdentity:
      type: object
      additionalProperties: false
      required: [id, username, display_name]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        username:
          type: string
        display_name:
          type: string
    ManagedUser:
      unevaluatedProperties: false
      allOf:
        - $ref: "#/components/schemas/UserIdentity"
        - type: object
          required:
            - status
            - failed_login_count
            - password_changed_at
            - created_at
            - updated_at
          properties:
            status:
              type: string
              enum: [active, locked, disabled]
            failed_login_count:
              type: integer
              minimum: 0
            locked_at:
              $ref: "#/components/schemas/Timestamp"
            lock_expires_at:
              $ref: "#/components/schemas/Timestamp"
            password_changed_at:
              $ref: "#/components/schemas/Timestamp"
            created_at:
              $ref: "#/components/schemas/Timestamp"
            updated_at:
              $ref: "#/components/schemas/Timestamp"
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, display_name, password]
      properties:
        username:
          type: string
        display_name:
          type: string
          maxLength: 253
        password:
          type: string
          format: password
          minLength: 15
          maxLength: 1024
    UpdateUserRequest:
      type: object
      additionalProperties: false
      required: [display_name]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 253
    SetUserStatusRequest:
      type: object
      additionalProperties: false
      required: [status, confirm]
      properties:
        status:
          type: string
          enum: [active, disabled]
        confirm:
          const: true
    Confirmation:
      type: object
      additionalProperties: false
      required: [confirm]
      properties:
        confirm:
          const: true
...
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [password, confirm]
      properties:
        password:
          type: string
          format: password
          minLength: 15
          maxLength: 1024
        confirm:
          const: true
    RoleBinding:
      type: object
      additionalProperties: false
      required: [id, subject_id, role, scope_type, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        subject_id:
          $ref: "#/components/schemas/UUID"
        role:
          type: string
          enum: [admin, viewer]
        scope_type:
          type: string
          enum: [global, tenant, project]
        tenant_id:
          $ref: "#/components/schemas/UUID"
        project_id:
          $ref: "#/components/schemas/UUID"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        replayed:
          type: boolean
    CreateRoleBindingRequest:
      type: object
      additionalProperties: false
      required: [subject_id, role, scope_type, confirm]
      properties:
        subject_id:
          $ref: "#/components/schemas/UUID"
        role:
          type: string
          enum: [admin, viewer]
        scope_type:
          type: string
          enum: [global, tenant, project]
        tenant_id:
          $ref: "#/components/schemas/UUID"
        project_id:
          $ref: "#/components/schemas/UUID"
        confirm:
          const: true
...
    Pagination:
      type: object
      additionalProperties: false
      required: [limit, offset, total, has_more]
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
        offset:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
        has_more:
          type: boolean
...